- move config to own file; - add code coverage ignore
