+++ .github/workflows/ci.yml
@@ -73,6 +73,10 @@ jobs:
       - name: "Test"
         run: |
           go test ./... -race -coverprofile=coverage.out -timeout 5m
+      - name: "Exclude codecoverage"
+        run: |
+          grep -v -E -f .covignore coverage.out > coverage.filtered.out
+          mv coverage.filtered.out coverage.out
       - name: "Code coverage"
         uses: codecov/codecov-action@v3
       - name: "Build artifacts"
