+++ internal/app/github/oauth/middleware.go
@@ -3,11 +3,9 @@ package oauth
 import (
 	"context"
 	"encoding/json"
-	"errors"
 	"fmt"
 	"log"
 	"net/http"
-	"os"
 	"path"
 	"strconv"
 	"strings"
@@ -21,13 +19,8 @@ import (
 	oauthgithub "golang.org/x/oauth2/github"
 )
 
-const ClientIdEnv = "GITHUBOAUTH_CLIENTID"
-const ClientSecretEnv = "GITHUBOAUTH_CLIENTSECRET"
-
 var (
-	cache        sync.Map
-	clientId     string = "set_by_init"
-	clientSecret string = "set_by_init"
+	cache sync.Map
 )
 
 func GithubOAuth() gin.HandlerFunc {
@@ -47,11 +40,6 @@ func GithubOAuth() gin.HandlerFunc {
 			ctx.Next()
 		}
 	}
-}
-
-func ReadConfig() (err error) {
-	clientId, clientSecret, err = loadOAuthConfig()
-	return err
 }
 
 func redirectToGithubOAuth(ctx *gin.Context) {
@@ -135,19 +123,4 @@ func (t *TokenSource) Token() (*oauth2.Token, error) {
 		AccessToken: t.AccessToken,
 	}
 	return token, nil
-}
-
-func loadOAuthConfig() (clientId string, clientSecret string, err error) {
-	clientId, found := os.LookupEnv(ClientIdEnv)
-	if !found {
-		err = errors.New("service host to generate documents is not configured. Expect environment variable " + ClientIdEnv)
-		return
-	}
-	clientSecret, found = os.LookupEnv(ClientSecretEnv)
-	if !found {
-		err = errors.New("service port to generate documents is not configured. Expect environment variable " + ClientSecretEnv)
-		return
-	}
-
-	return
 }
